Implement Edit Functionality and restring Delete only for the owner of the recipe.

+++ TestProjectForShareRecipies/Controllers/RecipeController.cs
@@ -1,5 +1,7 @@
 ﻿using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
+using Newtonsoft.Json;
+using System.Text;
 using TestProjectForShareRecipies.Infrastructure;
 using TestProjectForShareRecipies.Models.Recipe;
 using TestProjectForShareRecipies.Services.Recipe;
@@ -38,7 +40,7 @@ namespace TestProjectForShareRecipies.Controllers
         [HttpGet]
         public async Task<IActionResult> Details(int id)
         {
-            if (!(await recipeService.ExistRecipeAsync(id)))
+            if (!await recipeService.ExistRecipeAsync(id))
             {
                 return BadRequest();
             }
@@ -75,9 +77,71 @@ namespace TestProjectForShareRecipies.Controllers
         }
 
         [HttpGet]
-        public IActionResult EditRecipe(int id)
+        public async Task<IActionResult> Edit(int id)
         {
-            return View();
+            if(!await recipeService.ExistRecipeAsync(id))
+            {
+                return BadRequest();
+            }
+
+            if (!await recipeService.IsTheOwnerOfTheRecipe(id, User.Id()))
+            {
+                return Unauthorized();
+            }
+
+            var recipe = await recipeService.RecipeDetailsByIdAsync(id);
+            var builder = new StringBuilder();
+            List<string> ingredientsList = new List<string>();
+
+            foreach (var ingredient in recipe.Ingredients)
+            {
+                builder.Append(ingredient.Name);
+                builder.Append(" ");
+                builder.Append(ingredient.Quantity);
+                builder.Append(" ");
+                builder.Append(ingredient.MeassureUnit);
+                ingredientsList.Add(builder.ToString());
+                builder.Clear();
+            }
+            var ingredients = JsonConvert.SerializeObject(ingredientsList);
+
+            var recipeModel = new RecipeFormModel()
+            {
+                Name = recipe.Name,
+                Description = recipe.Description,
+                Picture = recipe.Picture,
+                AuthorId = User.Id(),
+                Ingredients = ingredients,
+                Categories = await recipeService.AllCategoriesAsync(),
+                MeassureUnits = await recipeService.AllMeassureUnitsAsync(),
+                CategoryId = (await recipeService.GetCategoryByNameAsync(recipe.Category)).Id
+            };
+
+            return View(recipeModel);
+        }
+
+        [HttpPost]
+        public async Task<IActionResult> Edit(RecipeFormModel model, int id)
+        {
+            if (!ModelState.IsValid)
+            {
+                return View();
+            }
+
+            if (!await recipeService.ExistRecipeAsync(id))
+            {
+                return BadRequest();
+            }
+
+            if (!await recipeService.IsTheOwnerOfTheRecipe(id, User.Id()))
+            {
+                return Unauthorized();
+            }
+
+            await recipeService.EditRecipeAsync(model, id);
+
+            return RedirectToAction(nameof(Details), new { id = id });
+
         }
 
         [HttpGet]
@@ -86,6 +150,11 @@ namespace TestProjectForShareRecipies.Controllers
             if (!await recipeService.ExistRecipeAsync(id))
             {
                 return BadRequest();
+            }
+
+            if (!await recipeService.IsTheOwnerOfTheRecipe(id, User.Id()))
+            {
+                return Unauthorized();
             }
 
             var model = await recipeService.RecipeDetailsByIdAsync(id);
@@ -99,6 +168,11 @@ namespace TestProjectForShareRecipies.Controllers
             if (!await recipeService.ExistRecipeAsync(id))
             {
                 return BadRequest();
+            }
+
+            if (!await recipeService.IsTheOwnerOfTheRecipe(id, User.Id()))
+            {
+                return Unauthorized();
             }
 
             await recipeService.DeleteAsync(id);
